onDragEnter and onDragLeave events. some cleanup.
