+++ src/utils/fileExtension.js
@@ -1,0 +1,10 @@
+const fileExtension = (file) => {
+   const extensionSplit = file.name.split('.')
+   if (extensionSplit.length > 1) {
+      return extensionSplit[extensionSplit.length - 1]
+   }
+
+   return 'none'
+}
+
+export default fileExtension
